Move methods to logic and dom files
- Update methods in dom to arrow functions
